backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
aiohttp>=3.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.26.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
